Implement Ring, go overboard
I don't like the way that Cowboy handles middleware.
Since a handler is responsible for calling the reply function,
it is difficult to add functionality to the stack that needs the
context of both the request and response (for example, applying
an operation to the response based on a value in the request,
like the Accepts header).

So, I've implemented Clojure's Ring architecture instead.